add support for legacy projects, fix migration failure

+++ source/database/migrations/2024_08_02_000000_create_earmark_hold_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 class CreateEarmarkHoldTable extends Migration
 {
@@ -13,15 +14,17 @@ class CreateEarmarkHoldTable extends Migration
      */
     public function up()
     {
-        Schema::create('earmark_hold', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer(config('earmark.columns.digit'));
-            $table->string(config('earmark.columns.group'))->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('earmark_hold')) {
+            Schema::create('earmark_hold', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer(config('earmark.columns.digit'));
+                $table->string(config('earmark.columns.group'))->nullable();
+                $table->timestamps();
 
-            $table->index(config('earmark.columns.digit'));
-            $table->index(config('earmark.columns.group'));
-        });
+                $table->index(config('earmark.columns.digit'));
+                $table->index(config('earmark.columns.group'));
+            });
+        }
 
         // Poing\Earmark\Models\EarMark::max('digit')
         // Poing\Earmark\Models\EarMark::where('type','alpha')->max('digit')
@@ -34,6 +37,9 @@ class CreateEarmarkHoldTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('earmark_hold');
+        $legacyMigration = DB::table('migrations')->where('migration', '2019_05_22_090330_create_earmark_hold_table')->first();
+        if(!$legacyMigration) {
+            Schema::dropIfExists('earmark_hold');
+        }
     }
 }
